cleanup all shaders and fix hot reload
